HyperThreading warning in hhl's help output
It has been observed that hhl performance may reduces when the number of
threads is greater than the number of real cores. In HyperThreading
setting the maximum number of theads suggested by OpenMP interface is
the number of virtual cores. Since this may cause a performance loss we
warn user about this in the program help.