Dummy parallel version of GHLp algorithm
GHLp has a sequential lazy-update queue, which is unlikely to be
parallelized. However since it is lazy queue we can utilize other cores
to recalculate values for non-top elements and update them to be more
relevant. Usually this gives us a speedup within the factor of 2
even when the number of threads is larger than 2.