Explain why we use vertex rank instead of id in HHL
When we construct HHL (Akiba, HHl and UHHL classes) we use vertex rank
instead of vertex id in hubs, e.g. when we add u to the label of v we
add (rank[u], dist(v,u)) to L(v). Note that the labeling is valid (rank
is a bijection).
HHL algorithms process vertices from the most important to the least
important, so the labels are ordered by hub rank and we don't need to
sort labels at the end.
However, this may lead to confusion if one is about to look at the
labels. Since we don't define labels in a formal way, let's just put a
comment near the labeling construction lines.